+++ src/bill/mod.rs
@@ -1,14 +1,10 @@
-use crate::blockchain::{Block, Chain, ChainToReturn, OperationCode};
+use crate::blockchain::{Chain, ChainToReturn, OperationCode};
 use crate::constants::{BILLS_FOLDER_PATH, BILLS_KEYS_FOLDER_PATH};
 use crate::external;
 use crate::service::contact_service::IdentityPublicData;
-use crate::util;
 use crate::util::file::is_not_hidden_or_directory;
 use borsh::BorshDeserialize;
 use borsh_derive::{BorshDeserialize, BorshSerialize};
-use identity::{get_whole_identity, read_peer_id_from_file};
-use openssl::pkey::Private;
-use openssl::rsa::Rsa;
 use rocket::serde::{Deserialize, Serialize};
 use rocket::FromForm;
 use std::fs;
@@ -235,399 +231,6 @@ pub async fn get_bills_for_list() -> Vec<BitcreditBillToReturn> {
         }
     }
     bills
-}
-
-pub async fn endorse_bitcredit_bill(
-    bill_name: &str,
-    endorsee: IdentityPublicData,
-    timestamp: i64,
-) -> bool {
-    let my_peer_id = read_peer_id_from_file().to_string();
-    let bill = read_bill_from_file(bill_name).await;
-
-    let mut blockchain_from_file = Chain::read_chain_from_file(bill_name);
-    let last_block = blockchain_from_file.get_latest_block();
-
-    let exist_block_with_code_endorse =
-        blockchain_from_file.exist_block_with_operation_code(OperationCode::Endorse);
-
-    let exist_block_with_code_mint =
-        blockchain_from_file.exist_block_with_operation_code(OperationCode::Mint);
-
-    let exist_block_with_code_sell =
-        blockchain_from_file.exist_block_with_operation_code(OperationCode::Sell);
-
-    if (my_peer_id.eq(&bill.payee.peer_id)
-        && !exist_block_with_code_endorse
-        && !exist_block_with_code_sell
-        && !exist_block_with_code_mint)
-        || (my_peer_id.eq(&bill.endorsee.peer_id))
-    {
-        let identity = get_whole_identity();
-
-        let my_identity_public =
-            IdentityPublicData::new(identity.identity.clone(), identity.peer_id.to_string());
-        let endorsed_by = serde_json::to_vec(&my_identity_public).unwrap();
-
-        let data_for_new_block_in_bytes = serde_json::to_vec(&endorsee).unwrap();
-        let data_for_new_block = "Endorsed to ".to_string()
-            + &hex::encode(data_for_new_block_in_bytes)
-            + " endorsed by "
-            + &hex::encode(endorsed_by);
-
-        let keys = read_keys_from_bill_file(bill_name);
-        let key: Rsa<Private> = Rsa::private_key_from_pem(keys.private_key_pem.as_bytes()).unwrap();
-
-        let data_for_new_block_in_bytes = data_for_new_block.as_bytes().to_vec();
-        let data_for_new_block_encrypted =
-            util::rsa::encrypt_bytes(&data_for_new_block_in_bytes, &key);
-        let data_for_new_block_encrypted_in_string_format =
-            hex::encode(data_for_new_block_encrypted);
-
-        let new_block = Block::new(
-            last_block.id + 1,
-            last_block.hash.clone(),
-            data_for_new_block_encrypted_in_string_format,
-            bill_name.to_owned(),
-            identity.identity.public_key_pem.clone(),
-            OperationCode::Endorse,
-            identity.identity.private_key_pem.clone(),
-            timestamp,
-        );
-
-        let try_add_block = blockchain_from_file.try_add_block(new_block.clone());
-        if try_add_block && blockchain_from_file.is_chain_valid() {
-            blockchain_from_file.write_chain_to_file(&bill.name);
-            true
-        } else {
-            false
-        }
-    } else {
-        false
-    }
-}
-
-pub async fn mint_bitcredit_bill(
-    bill_name: &str,
-    mintnode: IdentityPublicData,
-    timestamp: i64,
-) -> bool {
-    let my_peer_id = read_peer_id_from_file().to_string();
-    let bill = read_bill_from_file(bill_name).await;
-
-    let mut blockchain_from_file = Chain::read_chain_from_file(bill_name);
-    let last_block = blockchain_from_file.get_latest_block();
-
-    let exist_block_with_code_endorse =
-        blockchain_from_file.exist_block_with_operation_code(OperationCode::Endorse);
-
-    let exist_block_with_code_mint =
-        blockchain_from_file.exist_block_with_operation_code(OperationCode::Mint);
-
-    let exist_block_with_code_sell =
-        blockchain_from_file.exist_block_with_operation_code(OperationCode::Sell);
-
-    if (my_peer_id.eq(&bill.payee.peer_id)
-        && !exist_block_with_code_endorse
-        && !exist_block_with_code_sell
-        && !exist_block_with_code_mint)
-        || (my_peer_id.eq(&bill.endorsee.peer_id))
-    {
-        let identity = get_whole_identity();
-
-        let my_identity_public =
-            IdentityPublicData::new(identity.identity.clone(), identity.peer_id.to_string());
-        let minted_by = serde_json::to_vec(&my_identity_public).unwrap();
-
-        let data_for_new_block_in_bytes = serde_json::to_vec(&mintnode).unwrap();
-        let data_for_new_block = "Endorsed to ".to_string()
-            + &hex::encode(data_for_new_block_in_bytes)
-            + " endorsed by "
-            + &hex::encode(minted_by);
-
-        let keys = read_keys_from_bill_file(bill_name);
-        let key: Rsa<Private> = Rsa::private_key_from_pem(keys.private_key_pem.as_bytes()).unwrap();
-
-        let data_for_new_block_in_bytes = data_for_new_block.as_bytes().to_vec();
-        let data_for_new_block_encrypted =
-            util::rsa::encrypt_bytes(&data_for_new_block_in_bytes, &key);
-        let data_for_new_block_encrypted_in_string_format =
-            hex::encode(data_for_new_block_encrypted);
-
-        let new_block = Block::new(
-            last_block.id + 1,
-            last_block.hash.clone(),
-            data_for_new_block_encrypted_in_string_format,
-            bill_name.to_owned(),
-            identity.identity.public_key_pem.clone(),
-            OperationCode::Mint,
-            identity.identity.private_key_pem.clone(),
-            timestamp,
-        );
-
-        let try_add_block = blockchain_from_file.try_add_block(new_block.clone());
-
-        if try_add_block && blockchain_from_file.is_chain_valid() {
-            let bill_id = bill.name.clone();
-
-            blockchain_from_file.write_chain_to_file(&bill_id);
-            true
-        } else {
-            false
-        }
-    } else {
-        false
-    }
-}
-
-pub async fn sell_bitcredit_bill(
-    bill_name: &str,
-    buyer: IdentityPublicData,
-    timestamp: i64,
-    amount_numbers: u64,
-) -> bool {
-    let my_peer_id = read_peer_id_from_file().to_string();
-    let bill = read_bill_from_file(bill_name).await;
-
-    let mut blockchain_from_file = Chain::read_chain_from_file(bill_name);
-    let last_block = blockchain_from_file.get_latest_block();
-
-    let exist_block_with_code_endorse =
-        blockchain_from_file.exist_block_with_operation_code(OperationCode::Endorse);
-
-    let exist_block_with_code_sell =
-        blockchain_from_file.exist_block_with_operation_code(OperationCode::Sell);
-
-    if (my_peer_id.eq(&bill.payee.peer_id)
-        && !exist_block_with_code_endorse
-        && !exist_block_with_code_sell)
-        || (my_peer_id.eq(&bill.endorsee.peer_id))
-    {
-        let identity = get_whole_identity();
-
-        let my_identity_public =
-            IdentityPublicData::new(identity.identity.clone(), identity.peer_id.to_string());
-        let seller = serde_json::to_vec(&my_identity_public).unwrap();
-
-        let buyer_u8 = serde_json::to_vec(&buyer).unwrap();
-        let data_for_new_block = "Sold to ".to_string()
-            + &hex::encode(buyer_u8)
-            + " sold by "
-            + &hex::encode(seller)
-            + " amount: "
-            + &amount_numbers.to_string();
-
-        let keys = read_keys_from_bill_file(bill_name);
-        let key: Rsa<Private> = Rsa::private_key_from_pem(keys.private_key_pem.as_bytes()).unwrap();
-
-        let data_for_new_block_in_bytes = data_for_new_block.as_bytes().to_vec();
-        let data_for_new_block_encrypted =
-            util::rsa::encrypt_bytes(&data_for_new_block_in_bytes, &key);
-        let data_for_new_block_encrypted_in_string_format =
-            hex::encode(data_for_new_block_encrypted);
-
-        let new_block = Block::new(
-            last_block.id + 1,
-            last_block.hash.clone(),
-            data_for_new_block_encrypted_in_string_format,
-            bill_name.to_owned(),
-            identity.identity.public_key_pem.clone(),
-            OperationCode::Sell,
-            identity.identity.private_key_pem.clone(),
-            timestamp,
-        );
-
-        let try_add_block = blockchain_from_file.try_add_block(new_block.clone());
-        if try_add_block && blockchain_from_file.is_chain_valid() {
-            blockchain_from_file.write_chain_to_file(&bill.name);
-            true
-        } else {
-            false
-        }
-    } else {
-        false
-    }
-}
-
-pub async fn request_pay(bill_name: &str, timestamp: i64) -> bool {
-    let my_peer_id = read_peer_id_from_file().to_string();
-    let bill = read_bill_from_file(bill_name).await;
-
-    let mut blockchain_from_file = Chain::read_chain_from_file(bill_name);
-    let last_block = blockchain_from_file.get_latest_block();
-
-    let exist_block_with_code_endorse =
-        blockchain_from_file.exist_block_with_operation_code(OperationCode::Endorse);
-
-    let exist_block_with_code_mint =
-        blockchain_from_file.exist_block_with_operation_code(OperationCode::Mint);
-
-    let exist_block_with_code_sell =
-        blockchain_from_file.exist_block_with_operation_code(OperationCode::Sell);
-
-    if (my_peer_id.eq(&bill.payee.peer_id)
-        && !exist_block_with_code_endorse
-        && !exist_block_with_code_sell
-        && !exist_block_with_code_mint)
-        || (my_peer_id.eq(&bill.endorsee.peer_id))
-    {
-        let identity = get_whole_identity();
-
-        let my_identity_public =
-            IdentityPublicData::new(identity.identity.clone(), identity.peer_id.to_string());
-
-        let data_for_new_block_in_bytes = serde_json::to_vec(&my_identity_public).unwrap();
-        let data_for_new_block =
-            "Requested to pay by ".to_string() + &hex::encode(data_for_new_block_in_bytes);
-
-        let keys = read_keys_from_bill_file(bill_name);
-        let key: Rsa<Private> = Rsa::private_key_from_pem(keys.private_key_pem.as_bytes()).unwrap();
-
-        let data_for_new_block_in_bytes = data_for_new_block.as_bytes().to_vec();
-        let data_for_new_block_encrypted =
-            util::rsa::encrypt_bytes(&data_for_new_block_in_bytes, &key);
-        let data_for_new_block_encrypted_in_string_format =
-            hex::encode(data_for_new_block_encrypted);
-
-        let new_block = Block::new(
-            last_block.id + 1,
-            last_block.hash.clone(),
-            data_for_new_block_encrypted_in_string_format,
-            bill_name.to_owned(),
-            identity.identity.public_key_pem.clone(),
-            OperationCode::RequestToPay,
-            identity.identity.private_key_pem.clone(),
-            timestamp,
-        );
-
-        let try_add_block = blockchain_from_file.try_add_block(new_block.clone());
-        if try_add_block && blockchain_from_file.is_chain_valid() {
-            blockchain_from_file.write_chain_to_file(&bill.name);
-            true
-        } else {
-            false
-        }
-    } else {
-        false
-    }
-}
-
-pub async fn request_acceptance(bill_name: &str, timestamp: i64) -> bool {
-    let my_peer_id = read_peer_id_from_file().to_string();
-    let bill = read_bill_from_file(bill_name).await;
-
-    let mut blockchain_from_file = Chain::read_chain_from_file(bill_name);
-    let last_block = blockchain_from_file.get_latest_block();
-
-    let exist_block_with_code_endorse =
-        blockchain_from_file.exist_block_with_operation_code(OperationCode::Endorse);
-
-    let exist_block_with_code_sell =
-        blockchain_from_file.exist_block_with_operation_code(OperationCode::Sell);
-
-    let exist_block_with_code_mint =
-        blockchain_from_file.exist_block_with_operation_code(OperationCode::Mint);
-
-    if (my_peer_id.eq(&bill.payee.peer_id)
-        && !exist_block_with_code_endorse
-        && !exist_block_with_code_sell
-        && !exist_block_with_code_mint)
-        || (my_peer_id.eq(&bill.endorsee.peer_id))
-    {
-        let identity = get_whole_identity();
-
-        let my_identity_public =
-            IdentityPublicData::new(identity.identity.clone(), identity.peer_id.to_string());
-
-        let data_for_new_block_in_bytes = serde_json::to_vec(&my_identity_public).unwrap();
-        let data_for_new_block =
-            "Requested to accept by ".to_string() + &hex::encode(data_for_new_block_in_bytes);
-
-        let keys = read_keys_from_bill_file(bill_name);
-        let key: Rsa<Private> = Rsa::private_key_from_pem(keys.private_key_pem.as_bytes()).unwrap();
-
-        let data_for_new_block_in_bytes = data_for_new_block.as_bytes().to_vec();
-        let data_for_new_block_encrypted =
-            util::rsa::encrypt_bytes(&data_for_new_block_in_bytes, &key);
-        let data_for_new_block_encrypted_in_string_format =
-            hex::encode(data_for_new_block_encrypted);
-
-        let new_block = Block::new(
-            last_block.id + 1,
-            last_block.hash.clone(),
-            data_for_new_block_encrypted_in_string_format,
-            bill_name.to_owned(),
-            identity.identity.public_key_pem.clone(),
-            OperationCode::RequestToAccept,
-            identity.identity.private_key_pem.clone(),
-            timestamp,
-        );
-
-        let try_add_block = blockchain_from_file.try_add_block(new_block.clone());
-        if try_add_block && blockchain_from_file.is_chain_valid() {
-            blockchain_from_file.write_chain_to_file(&bill.name);
-            true
-        } else {
-            false
-        }
-    } else {
-        false
-    }
-}
-
-pub async fn accept_bill(bill_name: &str, timestamp: i64) -> bool {
-    let my_peer_id = read_peer_id_from_file().to_string();
-    let bill = read_bill_from_file(bill_name).await;
-
-    let mut blockchain_from_file = Chain::read_chain_from_file(bill_name);
-    let last_block = blockchain_from_file.get_latest_block();
-    let accepted = blockchain_from_file.exist_block_with_operation_code(OperationCode::Accept);
-
-    if bill.drawee.peer_id.eq(&my_peer_id) {
-        if !accepted {
-            let identity = get_whole_identity();
-
-            let my_identity_public =
-                IdentityPublicData::new(identity.identity.clone(), identity.peer_id.to_string());
-
-            let data_for_new_block_in_bytes = serde_json::to_vec(&my_identity_public).unwrap();
-            let data_for_new_block =
-                "Accepted by ".to_string() + &hex::encode(data_for_new_block_in_bytes);
-
-            let keys = read_keys_from_bill_file(bill_name);
-            let key: Rsa<Private> =
-                Rsa::private_key_from_pem(keys.private_key_pem.as_bytes()).unwrap();
-
-            let data_for_new_block_in_bytes = data_for_new_block.as_bytes().to_vec();
-            let data_for_new_block_encrypted =
-                util::rsa::encrypt_bytes(&data_for_new_block_in_bytes, &key);
-            let data_for_new_block_encrypted_in_string_format =
-                hex::encode(data_for_new_block_encrypted);
-
-            let new_block = Block::new(
-                last_block.id + 1,
-                last_block.hash.clone(),
-                data_for_new_block_encrypted_in_string_format,
-                bill_name.to_owned(),
-                identity.identity.public_key_pem.clone(),
-                OperationCode::Accept,
-                identity.identity.private_key_pem.clone(),
-                timestamp,
-            );
-
-            let try_add_block = blockchain_from_file.try_add_block(new_block.clone());
-            if try_add_block && blockchain_from_file.is_chain_valid() {
-                blockchain_from_file.write_chain_to_file(&bill.name);
-                true
-            } else {
-                false
-            }
-        } else {
-            false
-        }
-    } else {
-        false
-    }
 }
 
 async fn read_bill_with_chain_from_file(id: &str) -> BitcreditBillToReturn {
