+++ src/constants.rs
@@ -14,14 +14,11 @@ pub const MAX_FILE_NAME_CHARACTERS: usize = 50;
 pub const VALID_FILE_MIME_TYPES: [&str; 3] = ["image/jpeg", "image/png", "application/pdf"];
 
 // Paths
-pub const IDENTITY_FOLDER_PATH: &str = "identity";
 pub const BILLS_FOLDER_PATH: &str = "bills";
 pub const BILLS_KEYS_FOLDER_PATH: &str = "bills_keys";
 pub const QUOTES_MAP_FOLDER_PATH: &str = "quotes";
 pub const BOOTSTRAP_FOLDER_PATH: &str = "bootstrap";
-pub const IDENTITY_FILE_PATH: &str = "identity/identity";
 pub const IDENTITY_PEER_ID_FILE_PATH: &str = "identity/peer_id";
-pub const IDENTITY_ED_25529_KEYS_FILE_PATH: &str = "identity/ed25519_keys";
 pub const QUOTE_MAP_FILE_PATH: &str = "quotes/quotes";
 pub const BOOTSTRAP_NODES_FILE_PATH: &str = "bootstrap/bootstrap_nodes.json";
