+++ src/web/handlers/contacts.rs
@@ -1,12 +1,10 @@
 use super::super::data::{DeleteContactForm, EditContactForm, NewContactForm};
-use crate::constants::IDENTITY_FILE_PATH;
 use crate::service::contact_service::Contact;
 use crate::service::{self, Result, ServiceContext};
 use rocket::form::Form;
 use rocket::http::Status;
 use rocket::serde::json::Json;
 use rocket::{delete, get, post, put, State};
-use std::path::Path;
 
 #[get("/return")]
 pub async fn return_contacts(state: &State<ServiceContext>) -> Result<Json<Vec<Contact>>> {
@@ -19,15 +17,14 @@ pub async fn remove_contact(
     remove_contact_form: Form<DeleteContactForm>,
     state: &State<ServiceContext>,
 ) -> Result<Status> {
-    if !Path::new(IDENTITY_FILE_PATH).exists() {
-        Err(service::Error::PreconditionFailed)
-    } else {
-        state
-            .contact_service
-            .delete_identity_by_name(&remove_contact_form.name)
-            .await?;
-        Ok(Status::Ok)
+    if !state.identity_service.identity_exists().await {
+        return Err(service::Error::PreconditionFailed);
     }
+    state
+        .contact_service
+        .delete_identity_by_name(&remove_contact_form.name)
+        .await?;
+    Ok(Status::Ok)
 }
 
 #[post("/new", data = "<new_contact_form>")]
@@ -35,16 +32,15 @@ pub async fn new_contact(
     state: &State<ServiceContext>,
     new_contact_form: Form<NewContactForm>,
 ) -> Result<Json<Vec<Contact>>> {
-    if !Path::new(IDENTITY_FILE_PATH).exists() {
-        Err(service::Error::PreconditionFailed)
-    } else {
-        state
-            .contact_service
-            .add_peer_identity(&new_contact_form.name, &new_contact_form.node_id)
-            .await?;
-        let res = state.contact_service.get_contacts().await?;
-        Ok(Json(res))
+    if !state.identity_service.identity_exists().await {
+        return Err(service::Error::PreconditionFailed);
     }
+    state
+        .contact_service
+        .add_peer_identity(&new_contact_form.name, &new_contact_form.node_id)
+        .await?;
+    let res = state.contact_service.get_contacts().await?;
+    Ok(Json(res))
 }
 
 #[put("/edit", data = "<edit_contact_form>")]
@@ -52,14 +48,13 @@ pub async fn edit_contact(
     edit_contact_form: Form<EditContactForm>,
     state: &State<ServiceContext>,
 ) -> Result<Json<Vec<Contact>>> {
-    if !Path::new(IDENTITY_FILE_PATH).exists() {
-        Err(service::Error::PreconditionFailed)
-    } else {
-        state
-            .contact_service
-            .update_identity_name(&edit_contact_form.old_name, &edit_contact_form.name)
-            .await?;
-        let res = state.contact_service.get_contacts().await?;
-        Ok(Json(res))
+    if !state.identity_service.identity_exists().await {
+        return Err(service::Error::PreconditionFailed);
     }
+    state
+        .contact_service
+        .update_identity_name(&edit_contact_form.old_name, &edit_contact_form.name)
+        .await?;
+    let res = state.contact_service.get_contacts().await?;
+    Ok(Json(res))
 }
